In script new parameter {-mtn} is added which can have value true or false. True means move builds to var/www/{domain} if false then build will be created to project root folder set by option -b (build directory path). So if build directory path = artifacts. Then builds will be in /artifacts

Related work items: #440

+++ setup_domain_v1.sh
@@ -3,20 +3,23 @@
 # Set the domain name
 domain=$domain
 build_directory_name=$build_directory_name
+move_to_nginx=true
 
 # Define a usage function
 function usage {
 #  echo "Usage: $0 [-f|--filename FILENAME] [-s|--size SIZE]" >&2
   echo "Below is the list of accepted parameters"
-  echo "   -b, --builddirectoryname   Set the build directory name"
-  echo "   -d, --domain               Set domain"
-  echo "   -h, --help                 Display this help message"
+  echo "   -b,    --builddirectoryname   Set the build directory name"
+  echo "   -mtn   --movetonginx          If true then create virtual with base path (/var/www) (default: $move_to_nginx)"
+  echo "   -d,    --domain               Set domain"
+  echo "   -h,    --help                 Display this help message"
   exit 1
 }
 # Parse command-line options/parameters
 while [[ "$#" -gt 0 ]]; do
   case $1 in
     -b|--builddirectoryname) build_directory_name="$2"; shift ;;
+    -mtn|--movetonginx) move_to_nginx="$2"; shift ;;
     -d|--domain) domain="$2"; shift ;;
     -h|--help) usage ;;
     *) echo "Invalid option: $1" >&2; usage ;;
@@ -25,6 +28,10 @@ while [[ "$#" -gt 0 ]]; do
 done
 
 docroot=$( pwd )'/'"$build_directory_name"
+
+ if [ "$move_to_nginx" = "true" ]; then
+   docroot="/var/www/$domain"
+  fi
 
 # Create a new NGINX configuration file
 config="/etc/nginx/sites-available/$domain"
@@ -45,8 +52,10 @@ sudo sh -c "echo 'server {
 
 }' > $config"
 
-# Create a symbolic link to enable the site
-sudo ln -s $config /etc/nginx/sites-enabled/
+if [ ! -L "/etc/nginx/sites-available/$domain" ]; then
+  sudo ln -s $config /etc/nginx/sites-enabled/
+  echo "Symbolic link created at $link_path."
+fi
 
 
 # Test the configuration
